Plan 9 from Bell Labs Fourth Edition 2003-04-19

+++ sys/src/ape/lib/ap/stdio/vsnprintf.c
@@ -2,7 +2,7 @@
  * pANS stdio -- vsnprintf
  */
 #include "iolib.h"
-int vsnprintf(char *buf, int nbuf, const char *fmt, va_list args){
+int vsnprintf(char *buf, size_t nbuf, const char *fmt, va_list args){
 	int n;
 	char *v;
 	FILE *f=_IO_sopenw();
